Simple GoL logic implemented (seedTester.py)
A simple (and likely unoptimized) implementation of Conway's Game of
Life has been put together in seedTester.py. Details for using it are
included in the README; additionally, an example pattern has been
included ("patterns/testGlider.txt") as a demonstration of the GoL
Plaintext format.

On branch main

Changes to be committed:
	modified:   README.md
	new file:   patterns/testGlider.txt
	new file:   seedTester.py